minorfix

+++ Source/mechanics/mechanics.h
@@ -35,7 +35,7 @@ typedef struct {
 
 /* --- VARIABILI GLOBALI (EXTERN) --- */
 extern volatile uint8_t timer_tick;
-extern volatile uint16_t hardDrop_flag;
+extern volatile uint8_t hardDrop_flag;
 extern volatile uint16_t playing_field[HEIGHT][WIDTH];
 extern volatile uint32_t score;
 extern volatile uint32_t HighScore;
@@ -56,8 +56,8 @@ void movePieceRight(void);
 void movePieceDown(void);
 void handlePieceLock(void);
 void lockPiece(void);
-int deleteFullLines(void);
-int tryMoveDown(void);
+uint16_t deleteFullLines(void);
+uint8_t tryMoveDown(void);
 void hardDrop(void);
 void initializeGame(void);
 void initializePlayingField(void);
